% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_fun.R
\name{cat_fun}
\alias{cat_fun}
\title{Categorical summary}
\usage{
cat_fun(data, outcome, var, digits = 2)
}
\arguments{
\item{data}{data set with a two-level outcome and numerical/categorical covariates}

\item{outcome}{name of categorical grouping variable, must have two levels, string}

\item{var}{name of categorical covariate in the data set, may have any number of levels, string}

\item{digits}{for rounding on percents, default is 2}
}
\value{
A list of three items.
\itemize{
\item \code{cat_info} is a hash with keys = levels of \code{var} and values = another hash
with keys = levels of \code{outcome} and values = counts of \code{var} by \code{outcome}.
\item \code{para_pval} is the parametric p-value for test of differences of \code{var} using a
Chi-square test
\item \code{nonpara_pval} is a non-parametric p-value for test of differences of \code{var} using
Fisher's Exact test
}
}
\description{
\code{cat_fun()} will produce the components of contingency table
and a test of differences via parametric p-value (Chi-Square) and non-parametric p-value
(Fisher's Exact test). These summaries can be used in a crude preliminary Table 1 to share
with researchers in the beginning stages of analysis.
}
\details{
Summary statistics for a categorical covariate and a two-level outcome
}
\examples{
cat_fun(data=FEV, outcome="smoke", var="sex")
}
\author{
C. Rakowski
}
