% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEV.R
\docType{data}
\name{FEV}
\alias{FEV}
\title{Rosner's FEV data set}
\format{
A data frame with 654 observations and 6 variables:
\describe{
\item{age}{Age, in years}
\item{fev}{Maximum forced expiratory volume in one second}
\item{height}{Height in inches}
\item{id}{id for each row}
\item{sex}{'male' or 'female'}
\item{smoke}{'current smoker'  or 'non-current smoker'}
}
}
\source{
\url{https://hbiostat.org/data/}
}
\description{
A dataset containing data on respiratory status of smokers, single observation per id
}
