% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make table}
\usage{
make_table(
  data,
  outcome,
  num_vec = NULL,
  cat_vec = NULL,
  digits = 2,
  span = FALSE,
  total = FALSE
)
}
\arguments{
\item{data}{data set with a two-level outcome and numerical/categorical covariates}

\item{outcome}{categorical grouping variable, must have two levels, string}

\item{num_vec}{vector of names of numerical covariates (strings) in the data set}

\item{cat_vec}{vector of names of categorical covariates (strings) in the data set}

\item{digits}{number of digits to display, default is 2}

\item{span}{boolean, for whether to include StD for numeric; default is FALSE}

\item{total}{boolean, for whether to include an overall column; default is FALSE}
}
\value{
A table.
First two columns will have variables names, levels, and names of descriptive statistics.
Third and fourth columns will have the summary statistics for levels 1 and 2 of the outcome.
Fifth and sixth columns will have parametric and non-parametric p-values from a test of differences.
}
\description{
\code{make_table()} will create a rough summary table that should be readable enough
to share as a preliminary step when discussing with researchers.
Numerical variables will have mean, median, and counts while categorical variables will
have contingency tables. Two columns on the right will show the parametric and non-parametric
p-values for a test of differences. Parametric p-values will come from an ANOVA test for
numerical variables and a Chi-square test for categorical variables. Non-parametric p-values
will be from a Kruskal-Wallis test for numerical variables and Fisher's Exact test for
categorical variables.
}
\details{
Summary table for multiple covariates and a two-level outcome
}
\examples{
make_table(data=FEV, outcome="smoke", cat_vec="sex", num_vec=c("age", "height", "fev"))
}
\author{
C. Rakowski
}
