% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_fun.R
\name{numeric_fun}
\alias{numeric_fun}
\title{Numeric summary}
\arguments{
\item{data}{data set}

\item{outcome}{categorical grouping variable}

\item{var}{name of numerical covariate in the data set}
}
\value{
vector of sample size, mean, median, and p-value for test of differences (nonpar and par)
}
\description{
Description: To produce summary stats for a numeric variable
}
\details{
We want to produce sample size, mean, median, and test of differences
}
\author{
C. Rakowski
}
